backend/requirements.txt
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
mysqlclient==2.2.0
boto3==1.34.0
python-dotenv==1.0.0
gunicorn==21.2.0